Refactored a bit, prepared to place an abstraction over the Kafka implementation (so we do not depend on Kafka for local execution)
