+++ src/main/scala/org/codefeedr/Library/Internal/Kafka/KafkaController.scala
@@ -1,4 +1,4 @@
-package org.codefeedr.Library.Internal
+package org.codefeedr.Library.Internal.Kafka
 
 import org.apache.kafka.clients.admin.{AdminClient, NewTopic}
 import resource.managed
